+++ packages/deployer/internal/router-helper.js
@@ -1,3 +1,7 @@
+const fs = require('fs');
+const path = require('path');
+const relativePath = require('@synthetixio/core-js/utils/relative-path');
+
 /**
  * Converts the contracts name to private _CONSTANT_CASE format.
  * E.g.:
@@ -9,6 +13,17 @@ function toPrivateConstantCase(name) {
   return name.replace(/(?<![A-Z])[A-Z]/g, '_$&').toUpperCase();
 }
 
+function getRouterSource() {
+  const routerPath = path.join(
+    relativePath(hre.config.paths.sources, hre.config.paths.root),
+    'Router.sol'
+  );
+
+  const source = fs.readFileSync(routerPath).toString();
+  return source;
+}
+
 module.exports = {
   toPrivateConstantCase,
+  getRouterSource,
 };
