+++ packages/deployer/internal/router-source-validator.js
@@ -1,61 +1,48 @@
-const fs = require('fs');
-const path = require('path');
-const relativePath = require('@synthetixio/core-js/utils/relative-path');
 const { getModulesSelectors } = require('./contract-helper');
+const { getRouterSource } = require('./router-helper');
 
-function getRouterSource() {
-  const routerPath = path.join(
-    relativePath(hre.config.paths.sources, hre.config.paths.root),
-    'Router.sol'
-  );
+class RouterSourceValidator {
+  async findRepeatedModuleSelectors() {
+    const moduleSelectors = await getModulesSelectors();
+    const source = getRouterSource();
 
-  const source = fs.readFileSync(routerPath).toString();
-  return source;
-}
+    const errors = [];
+    moduleSelectors.forEach((moduleSelector) => {
+      const regex = `(:?\\s|^)case ${moduleSelector.selector}\\s.+`;
+      const matches = source.match(new RegExp(regex, 'gm'));
 
-async function findRepeatedSelectorsInSource() {
-  const moduleSelectors = await getModulesSelectors();
-  const source = getRouterSource();
+      if (matches && matches.length > 1) {
+        if (!errors.some((value) => value.selector === moduleSelector.selector)) {
+          errors.push({
+            msg: `Selector case ${moduleSelector.selector} found ${matches.length} times instead of the expected single time. Matches: ${matches}`,
+            repeatedInRouter: true,
+            selector: moduleSelector.selector,
+          });
+        }
+      }
+    });
+    return errors;
+  }
 
-  const errors = [];
-  moduleSelectors.forEach((moduleSelector) => {
-    const regex = `(:?\\s|^)case ${moduleSelector.selector}\\s.+`;
-    const matches = source.match(new RegExp(regex, 'gm'));
+  async findMissingModuleSelectors() {
+    const moduleSelectors = await getModulesSelectors();
+    const source = getRouterSource();
 
-    if (matches && matches.length > 1) {
-      if (!errors.some((value) => value.selector === moduleSelector.selector)) {
+    const errors = [];
+    moduleSelectors.forEach((moduleSelector) => {
+      const regex = `(:?\\s|^)case ${moduleSelector.selector}\\s.+`;
+      const matches = source.match(new RegExp(regex, 'gm'));
+
+      if (!matches || matches.length < 1) {
         errors.push({
-          msg: `Selector case ${moduleSelector.selector} found ${matches.length} times instead of the expected single time. Matches: ${matches}`,
-          repeatedInRouter: true,
-          selector: moduleSelector.selector,
+          msg: `Selector for ${moduleSelector.contractName}.${moduleSelector.name} not found in router`,
+          moduleSelector,
+          missingInRouter: true,
         });
       }
-    }
-  });
-  return errors;
+    });
+    return errors;
+  }
 }
 
-async function findMissingSelectorsInSource() {
-  const moduleSelectors = await getModulesSelectors();
-  const source = getRouterSource();
-
-  const errors = [];
-  moduleSelectors.forEach((moduleSelector) => {
-    const regex = `(:?\\s|^)case ${moduleSelector.selector}\\s.+`;
-    const matches = source.match(new RegExp(regex, 'gm'));
-
-    if (!matches || matches.length < 1) {
-      errors.push({
-        msg: `Selector for ${moduleSelector.contractName}.${moduleSelector.name} not found in router`,
-        moduleSelector,
-        missingInRouter: true,
-      });
-    }
-  });
-  return errors;
-}
-
-module.exports = {
-  findMissingSelectorsInSource,
-  findRepeatedSelectorsInSource,
-};
+module.exports = RouterSourceValidator;
