+++ packages/deployer/subtasks/validate-router.js
@@ -1,17 +1,8 @@
 const { subtask } = require('hardhat/config');
 const logger = require('@synthetixio/core-js/utils/logger');
-const { getSourcesAST } = require('../internal/ast/ast-sources');
-const {
-  findDuplicateSelectorsInAST,
-  findMissingSelectorsInAST,
-  findUnreachableSelectorsInAST,
-} = require('../internal/ast/router-ast-validator');
-const {
-  findMissingSelectorsInSource,
-  findRepeatedSelectorsInSource,
-} = require('../internal/router-source-validator');
-const filterValues = require('filter-values');
-
+const { getAllContractASTs } = require('@synthetixio/core-js/utils/hardhat');
+const RouterSourceValidator = require('../internal/router-source-validator');
+const RouterASTValidator = require('../internal/router-ast-validator');
 const { SUBTASK_VALIDATE_ROUTER, SUBTASK_CANCEL_DEPLOYMENT } = require('../task-names');
 
 subtask(
@@ -20,36 +11,50 @@ subtask(
 ).setAction(async (_, hre) => {
   logger.subtitle('Validating router');
 
-  const { contracts } = await getSourcesAST(hre);
-  const modules = filterValues(hre.deployer.deployment.general.contracts, (c) => c.isModule);
-
-  let sourceErrorsFound = [];
-  let astErrorsFound = [];
-  // Source Code checking
-  logger.debug('Validating Router source code');
-  sourceErrorsFound.push(...(await findMissingSelectorsInSource()));
-  sourceErrorsFound.push(...(await findRepeatedSelectorsInSource()));
-  if (sourceErrorsFound.length > 0) {
-    sourceErrorsFound.forEach((error) => {
-      logger.error(error.msg);
-    });
-  }
-
-  // AST checking
-  logger.debug('Validating Router compiled code');
-  astErrorsFound.push(...(await findMissingSelectorsInAST(contracts)));
-  astErrorsFound.push(...(await findUnreachableSelectorsInAST(contracts, modules)));
-  astErrorsFound.push(...(await findDuplicateSelectorsInAST(contracts)));
-  if (astErrorsFound.length > 0) {
-    astErrorsFound.forEach((error) => {
-      logger.error(error.msg);
-    });
-  }
+  const sourceErrorsFound = await _runSourceValidations();
+  const astErrorsFound = await _runASTValidations();
 
   if (sourceErrorsFound.length > 0 || astErrorsFound.length > 0) {
     logger.error('Router is not valid');
+
     return await hre.run(SUBTASK_CANCEL_DEPLOYMENT);
   }
 
   logger.checked('Router is valid');
 });
+
+async function _runSourceValidations() {
+  const errorsFound = [];
+
+  const validator = new RouterSourceValidator();
+
+  logger.debug('Validating Router source code');
+  errorsFound.push(...(await validator.findMissingModuleSelectors()));
+  errorsFound.push(...(await validator.findRepeatedModuleSelectors()));
+  if (errorsFound.length > 0) {
+    errorsFound.forEach((error) => {
+      logger.error(error.msg);
+    });
+  }
+
+  return errorsFound;
+}
+
+async function _runASTValidations() {
+  const errorsFound = [];
+
+  const asts = await getAllContractASTs(hre);
+  const validator = new RouterASTValidator(asts);
+
+  logger.debug('Validating Router compiled code');
+  errorsFound.push(...(await validator.findMissingModuleSelectors()));
+  errorsFound.push(...(await validator.findUnreachableModuleSelectors()));
+  errorsFound.push(...(await validator.findDuplicateModuleSelectors()));
+  if (errorsFound.length > 0) {
+    errorsFound.forEach((error) => {
+      logger.error(error.msg);
+    });
+  }
+
+  return errorsFound;
+}
